# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Faarooq Backend API",
    "version": "0.1.0",
    "description": "OpenAPI spec for the Twitter-like backend (Express + Supabase)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" },
    { "url": "https://faarooq-backend.onrender.com", "description": "Production" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "responses": {
      "RateLimited": {
        "description": "Too many requests",
        "headers": {
          "Retry-After": { "schema": { "type": "integer" }, "description": "Seconds until reset" },
          "X-RateLimit-Limit": { "schema": { "type": "integer" } },
          "X-RateLimit-Remaining": { "schema": { "type": "integer" } },
          "X-RateLimit-Reset": {
            "schema": { "type": "integer" },
            "description": "Unix epoch seconds when window resets"
          }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "Not found",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            },
            "required": ["code"]
          }
        },
        "required": ["error"]
      },
...
    }
  },
  "paths": {
...
    "/api/posts": {
      "get": {
        "summary": "Get timeline feed (posts + repost activities; excludes soft-deleted posts)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "cursor", "in": "query", "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK" },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "post": {
        "summary": "Create post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "text": { "type": "string", "maxLength": 280 } },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "summary": "Get single post (respects soft-delete & block/mute)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/reposts": {
      "post": {
        "summary": "Repost a post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "post_id": { "type": "string", "format": "uuid" } },
                "required": ["post_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" }
        }
      },
      "delete": {
        "summary": "Unrepost a post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "post_id": { "type": "string", "format": "uuid" } },
                "required": ["post_id"]
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/posts/{id}/reply": {
      "post": {
        "summary": "Reply to a post",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "text": { "type": "string", "maxLength": 280 } },
                "required": ["text"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/posts/{id}/replies": {
      "get": {
        "summary": "List replies for a post (excludes soft-deleted)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}